pyproject.toml
[project]
name = "chantiko-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlmodel",
    "litellm",
    "python-jose[cryptography]",
    "httpx",
    "psycopg2-binary",
    "mangum",
]
